Move implementation files from src to examples
