+++ examples/implementation/JsonStorage.ts
