+++ test/ContentTypeCache.test.ts
@@ -3,8 +3,8 @@ import axiosMockAdapter from 'axios-mock-adapter';
 import mockdate from 'mockdate';
 
 import ContentTypeCache from '../src/ContentTypeCache';
-import EditorConfig from '../src/EditorConfig';
-import InMemoryStorage from '../src/InMemoryStorage';
+import EditorConfig from '../examples/implementation/EditorConfig';
+import InMemoryStorage from '../examples/implementation/InMemoryStorage';
 
 const axiosMock = new axiosMockAdapter(axios);
