+++ test/PackageExporter.test.ts
@@ -4,14 +4,14 @@ import { withDir, withFile } from 'tmp-promise';
 import yauzlPromise from 'yauzl-promise';
 
 import ContentManager from '../src/ContentManager';
-import EditorConfig from '../src/EditorConfig';
-import FileContentStorage from '../src/FileContentStorage';
-import FileLibraryStorage from '../src/FileLibraryStorage';
+import EditorConfig from '../examples/implementation/EditorConfig';
+import FileContentStorage from '../examples/implementation/FileContentStorage';
+import FileLibraryStorage from '../examples/implementation/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import PackageExporter from '../src/PackageExporter';
 import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
-import User from '../src/User';
+import User from '../examples/implementation/User';
 
 describe('package exporter', () => {
     async function importAndExportPackage(packagePath: string): Promise<void> {
