+++ test/PackageImporter.test.ts
@@ -5,13 +5,13 @@ import { BufferWritableMock } from 'stream-mock';
 import { withDir } from 'tmp-promise';
 
 import ContentManager from '../src/ContentManager';
-import EditorConfig from '../src/EditorConfig';
-import FileContentStorage from '../src/FileContentStorage';
-import FileLibraryStorage from '../src/FileLibraryStorage';
+import EditorConfig from '../examples/implementation/EditorConfig';
+import FileContentStorage from '../examples/implementation/FileContentStorage';
+import FileLibraryStorage from '../examples/implementation/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
 import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
-import User from '../src/User';
+import User from '../examples/implementation/User';
 
 describe('package importer', () => {
     it('installs libraries', async () => {
