+++ test/PackageValidator.test.ts
@@ -1,7 +1,7 @@
 import * as fsExtra from 'fs-extra';
 import * as path from 'path';
 
-import EditorConfig from '../src/EditorConfig';
+import EditorConfig from '../examples/implementation/EditorConfig';
 import PackageValidator from '../src/PackageValidator';
 import TranslationService from '../src/TranslationService';
